requirements.txt
geopandas
google-cloud-bigquery
google-cloud-storage
matplotlib
matplotlib_scalebar
pandas
pandas_gbq
pydantic==1.10.11
shapely
streamlit
streamlit_pydantic
